feat: New Calculator class
